Changes to be able to operate with the new structure for the duty cycle. Only a single duty cycle is going to be plotted at a time
.txt and h5 files updated for the 10 materials.
Excitation "Triangle" renamed as "Triangular" for consistency.
Margin for the duty cycle shouldn't be changed so far, set as constant with a value of 0.01.
Triangular data is now read directly from the trapezoidal H5 files.
Duty cycle selector changed to sliders, so there is a single value selected each time for Triangular.
Function for the header modified, displaying mT and Hz for B and f respectively.
Bmin and Bmax modified to T.
Datasheet data cannot be displayed yet, as its format is not compatible with the Sinusoidal or Trapezoidal data. I am working on the new datasheet files, but it will take some time.
Since the same plot is plotted for Sinusoidal, Triangular and Trapezoidal, the code for these 3 parts have been somewhat merged.
This version works, although the code is not yet nice, I will be working on it too.