H5 files updated, txt files deleted, and some details
The JSON files have been updated to contain the info previously contained by the .txt files.
JSON to H5 updated with the same purpose.
txt files and "from importlib.resources import path" is no longer required (both in the ui_db.py and ui_raw.py files), a new function is added to read the metadata.
Modifications to the plots, some datasheet files added to test the operation, but most of them are not generated yet.
Titles of the plots modified.