Limit to the outlier factor added and minor changes
Some errors fixed.
Duty cycle variables based on c.streamlit, we have to separate the variables used for the core loss database and the core loss analysis.
Outlier function plot with the same axis as the plot in the right.
Some data has very large outlier factor, a function is added to delete the outlier data to the desired degree. If it is too complicated, we can select a constant value instead of offering a slider.