Improvements in Core Loss Database
Structure of the code modified, size of the columns depending on whether there is data to show or not, for each columns, different plots are shown depending on the type of excitation.
Waveform visualization with time in the x axis added, for the average flux and frequency in the slider.
Placeholder for DC flux, temperature, and some datasheet  H5 files added (empty) to avoid getting error messages. I will work on the datasheet files later on.