Adding the single cycle zip to the download function
The download page now has two download buttons, one for the raw files, the other for the B-H files with 100 samples.
It worked when I changed the path to my local one, and I hope it works with the files in the server too, but I don't know how to check that before committing.