Matlab Script with example
The Matlab script used to generate all the files has been uploaded to the Script folder.
This script has been modified so the user only have to modify the files in the Input folder, run the script, and the output files will be generated in the Output folder.
As an example, a reduced input dataset for N87 is provided (please note that since the original .csv files have been modified, time variables have been truncated so the sampling time is not perfect, but this is just an example of how to use it).
This script only works for our data so far, and constant sampling time, it is highly likely that for some other inputs, the scripts has to be modified. So it is only intended as a starting point for people developing their own tools rather than a tool that can be used in all scenarios.