Adding info to the intro page
There is no nice way to operate with dataframes when trying to display columns with different formats.
I haven't found a nice solution to add the manufacturer logo in the right of the table, so basically distances are adjusted manually, which means that when adding a new material, this part has to be modified too.