T and Hdc added, interpolation and outlier factor removed
Version with the temperature and dc bias measurement (new data).
Since we have sinusoidal data at different temperatures the interpolated datasheet information is no longer needed, making the addition of new materials much easier.
The data interpolation is removed, as interpolating for these sizes and dimensions is problematic.
The outlier factor is removed, as a significant contribution is because of "rounded" variables (D,Hdc) used to report the data, plus, it makes adding external data complicated as it is a parameter that needs to be in the dataframe.
Since interpolated data is removed, it is also removed from the comparison to iGSE and ML for sinusoidal data.